resolves #486 add test;
